Renamed "external" package to "fq" since it more clearly describe intention.
